refactor(daemon): move writer epoch authority out of fleet
