+++ tools/write-road-registry.json
@@ -38,7 +38,7 @@
     "packages/daemon/src/doc-sync-candidate-scanner.ts",
     "packages/daemon/src/fleet/center.ts",
     "packages/daemon/src/fleet/edge.ts", "packages/daemon/src/lease-broker.ts",
-    "packages/daemon/src/fleet/writer-epoch.ts",
+    "packages/daemon/src/writer-epoch.ts",
     "packages/daemon/src/fleet/replica-ack-store.ts",
     "packages/daemon/src/fleet/replica-cut-store.ts",
     "packages/daemon/src/fleet-edge-mirror.ts",
